face-service/requirements.txt
fastapi
uvicorn
deepface
pymongo
numpy
python-multipart
dnspython
python-dotenv
tensorflow
